feat: reorganizing folders and files
