+++ mental_health/mental_health.py
@@ -2,7 +2,7 @@
 import reflex as rx
 
 from . import pages
-from . import navigation
+from .states import navigation
 
 app = rx.App(
     theme=rx.theme(
@@ -16,7 +16,7 @@ app = rx.App(
 )
 
 #Website Pages
-app.add_page(pages.homepage, navigation.routes.HOME_ROUTE, title="Homepage")
-app.add_page(pages.chatpage, navigation.routes.CHAT_ROUTE, title="Chat")
+app.add_page(pages.homepage, navigation.HOME_ROUTE, title="Homepage")
+app.add_page(pages.chatpage, navigation.CHAT_ROUTE, title="Chat")
