--- mental_health/navigation/__init__.py
@@ -1,8 +1,0 @@
-from . import routes
-from .state import NavState
-
-
-__all__ = [
-    'routes',
-    'NavState'
-]
