--- mental_health/navigation/routes.py
@@ -1,2 +1,0 @@
-HOME_ROUTE="/"
-CHAT_ROUTE="/chatpage"
