+++ mental_health/states/__init__.py
@@ -1,0 +1,5 @@
+from .navigation import NavState
+
+__all__ = [
+    'NavState'
+]
